requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
bcrypt==4.1.1
python-jose==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0

# PDF & Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pillow==10.1.0

# AI & ML Libraries
langchain==0.1.0
openai==1.3.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Text Processing
nltk==3.8.1
textstat==0.7.3

# Language Translation
deep-translator==1.11.4

# PDF Generation
reportlab==4.0.9

# Database
psycopg2-binary==2.9.9

# API & Utilities
requests==2.31.0
aiofiles==23.2.1

# Security & JWT
cryptography==41.0.7
passlib==1.7.4

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
